Split the code. Introduce `engine` and `wasm` directories which represents engine of the game and wasm provider which might be used on Web

+++ src/engine/cell.rs
@@ -1,4 +1,4 @@
-pub type CellId = u32;
+pub type CellId = u16;
 
 /// Should contain 2 structure
 ///  1. Factory - to create Minesweeper engine
